[del/#203] 스케쥴 관련 코드 제거
