+++ app/src/main/java/com/bongtu/baekseo/presentation/record/RecordScreen.kt
@@ -25,8 +25,6 @@ import androidx.lifecycle.flowWithLifecycle
 import com.bongtu.baekseo.core.common.state.UiState
 import com.bongtu.baekseo.core.common.type.AttendType
 import com.bongtu.baekseo.core.common.type.EventCategoryType
-import com.bongtu.baekseo.core.common.type.ScheduleType
-import com.bongtu.baekseo.core.designsystem.component.BongBaekScheduleEmptyContent
 import com.bongtu.baekseo.core.designsystem.component.topbar.BongBaekCategoryBar
 import com.bongtu.baekseo.core.designsystem.theme.BongBaekTheme
 import com.bongtu.baekseo.data.model.event.ScheduleEvent
@@ -34,6 +32,7 @@ import com.bongtu.baekseo.presentation.record.RecordContract.RecordSideEffect.Na
 import com.bongtu.baekseo.presentation.record.RecordContract.RecordSideEffect.NavigateToDetail
 import com.bongtu.baekseo.presentation.record.RecordContract.RecordUiState
 import com.bongtu.baekseo.presentation.record.component.AttendTypeTab
+import com.bongtu.baekseo.presentation.record.component.BongBaekScheduleEmptyContent
 import com.bongtu.baekseo.presentation.record.component.RecordListContent
 import com.bongtu.baekseo.presentation.record.component.RecordTopBar
 import kotlinx.collections.immutable.persistentListOf
@@ -154,7 +153,6 @@ private fun RecordScreen(
                     isEnterDeleteButtonVisible = false
                     BongBaekScheduleEmptyContent(
                         eventType = category.label,
-                        scheduleType = ScheduleType.RECORD,
                         onButtonClick = navigateToAdd,
                         modifier = Modifier
                             .padding(
